Chimera_Pro_Live/requirements.txt
pandas
numpy
yfinance
streamlit
plotly
scipy
smartapi-python
pyotp
